component: add decorator variants for decorator components
so that we are able to put something like:

Button(...) | Maybe(&show_button)

* add operator|() which is a tiny shim for gluing decorator
  style functions with the decorated component. so we can
  chain decorators to wrap the decorated component to form
  the expected component.
* add decorator overrides for CatchEvent(), Maybe()
* update examples for them.

we cannot add the override for

Component Renderer(Component child, std::function<Element()>)

because otherwise we'd have a

ComponentDecorator Renderer(std::function<Element()>);

which is ambiguous with

Component Renderer(std::function<Element()>);

as their only difference is the return type, and C++ cannot deduce
the expected called function by looking at the return type. it does
the argument-dependent lookup instead.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>