Improve unicode codepoint Bisearch performance
Improve the performance of the functions for searching for codepoints
in a table by passing the table array in as a reference instead of copying
it.

This doesn't really improve the results of the benchmarks for some reason,
but in my application with GCC8 i686, Intel Atom E3950 and a 180x60 terminal,
this reduces the time for rendering a frame from ~31 ms to ~5.5 ms.